# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Card Management API
  description: |
    REST API для управления банковскими картами,
    пользователями, запросами на блокировку
    и переводами между собственными картами.
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/admin/users/{userId}/cards:
    post:
      tags:
      - Карты
      summary: Создать карту
      description: |
        Создаёт новую банковскую карту
        для указанного пользователя.
      operationId: createCard
      parameters:
      - name: userId
        in: path
        description: Идентификатор владельца карты
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "201":
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
  /api/admin/cards/{cardId}/block:
    patch:
      tags:
      - Карты
      summary: Заблокировать карту
      description: Блокирует выбранную банковскую карту
      operationId: blockCard
      parameters:
      - name: cardId
        in: path
        description: Идентификатор карты
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Карта успешно заблокирована
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: |
            Карта уже заблокирована
            или срок её действия истёк
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
  /api/admin/cards/{cardId}/activate:
    patch:
      tags:
      - Карты
      summary: Активировать карту
      description: Активирует выбранную банковскую карту
      operationId: activateCard
      parameters:
      - name: cardId
        in: path
        description: Идентификатор карты
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Карта успешно активирована
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: |
            Карта уже активна
            или срок её действия истёк
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
...
  /api/cards:
    get:
      tags:
      - Карты
      summary: Получить свои карты
      description: |
        Возвращает страницу банковских карт,
        принадлежащих текущему пользователю.

        Поддерживает поиск по последним четырём
        цифрам номера карты, фильтрацию по статусу,
        пагинацию и сортировку.
      operationId: getMyCards
      parameters:
      - name: lastFour
        in: query
        description: |
          Последние четыре цифры номера карты.
          Должны состоять ровно из четырёх цифр.
        required: false
        schema:
          type: string
        example: 3456
      - name: status
        in: query
        description: Статус карты
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        example: ACTIVE
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: Страница карт успешно получена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelCardResponse"
        "400":
          description: |
            Некорректные последние четыре цифры
            или значение статуса карты
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
  /api/cards/{cardId}:
    get:
      tags:
      - Карты
      summary: Получить свою карту
      description: |
        Возвращает карту только в том случае,
        если она принадлежит текущему пользователю.
      operationId: getMyCard
      parameters:
      - name: cardId
        in: path
        description: Идентификатор карты
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Карта успешно получена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Нет доступа к чужой карте
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
  /api/cards/{cardId}/balance:
    get:
      tags:
      - Карты
      summary: Получить баланс своей карты
      description: |
        Возвращает баланс карты только в том случае,
        если она принадлежит текущему пользователю.
      operationId: getMyCardBalance
      parameters:
      - name: cardId
        in: path
        description: Идентификатор карты
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Баланс успешно получен
          content:
            '*/*':
              schema:
                type: number
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Нет доступа к чужой карте
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
  /api/admin/cards:
    get:
      tags:
      - Карты
      summary: Получить все карты
      description: Возвращает страницу всех банковских карт
      operationId: getAllCards
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: Страница карт успешно получена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelCardResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      description: Единый формат ошибки REST API
      properties:
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: 2026-07-28T19:34:52.426
        status:
          type: integer
          format: int32
          description: HTTP-статус ошибки
          example: 404
        error:
          type: string
          description: Наименование HTTP-ошибки
          example: Not Found
        message:
          type: string
          description: Сообщение для пользователя
          example: Карта с идентификатором 15 не найдена
        path:
          type: string
          description: Путь HTTP-запроса
          example: /api/cards/15
...
    CardResponse:
      type: object
      description: Информация о банковской карте
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор карты
          example: 10
        userId:
          type: integer
          format: int64
          description: Идентификатор владельца карты
          example: 1
        maskedNumber:
          type: string
          description: Маскированный номер карты
          example: '**** **** **** 1234'
        expirationDate:
          type: string
          format: date
          description: Срок действия карты
          example: 2031-07-31
        status:
          type: string
          description: Статус карты
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
          example: ACTIVE
        balance:
          type: number
          description: Баланс карты
          example: 12500.5
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModelCardResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardResponse"
        page:
          $ref: "#/components/schemas/PageMetadata"
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT access-токен для авторизации
      scheme: bearer
      bearerFormat: JWT